py_client/requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.1
ddtrace==0.57.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
grpcio==1.46.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.7.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.19.0
pymongo==3.12.0
pyparsing==2.4.7
requests==2.27.1
six==1.16.0
tenacity==8.0.1
typing-extensions==3.10.0.0
urllib3==1.26.8
Werkzeug==2.0.1
zipp==3.5.0
